ME 135 Labs
ME 135 (Design of Microprocessor-Based Mechanical Systems) provides preparation for the conceptual design and prototyping of mechanical systems that use microprocessors to control machine activities, acquire and analyze data, and interact with operators. The architecture of microprocessors is related to problems in mechanical systems through study of systems, including electro-mechanical components, thermal components and a variety of instruments. Laboratory exercises lead through studies of different levels of software.